+++ Adventure_Game/BigRat.java
@@ -1,0 +1,74 @@
+package monster;
+
+public class BigRat extends Monster {
+    public BigRat() {
+        super(3, """
+                The leader of the pack is a truly monstrous rat, far larger than the rest, standing nearly three times their size. 
+                It towers over the others, as large as a medium-sized dog, with a hulking, muscular frame that ripples under its patchy, mangy fur. 
+                The fur, matted with filth, blood, and grime, clings in tufts to its hide, leaving exposed patches of rough, scaly skin that glisten with a sickly sheen. 
+                The rat\'s hide is battle-scarred, riddled with old wounds and deep gashes from past fights, evidence of its dominance and long survival in a brutal world. 
+                Its head is enormous, broad and flat, with jagged, discolored teeth that protrude from its mouth like broken shards of bone. 
+                These teeth are yellowed and cracked, but incredibly sharp, grinding together with an audible snap as the rat gnashes its jaws. 
+                Two long, curled fangs jut downward from its upper jaw, dripping with foamy saliva. 
+                
+                Its maw is perpetually open, revealing a cavernous mouth that can easily tear through flesh and bone. 
+                The foam is thick and frothy, streaked with blood from previous kills. 
+                The rat\'s eyes glow an unnatural red, bulging from its skull as if on the verge of popping out. 
+                Its pupils contracted into thin slits, twitching and flicking in every direction, never resting. 
+                Unlike the smaller rats, whose movements are jerky and erratic, this beast moves with a terrifying purpose. 
+                Every step is heavy and deliberate, its claws digging deep into the ground as it strides forward, scattering the smaller rats in its wake. 
+                """, """
+                It lunges forward, jaws snapping with a feral snarl. 
+                You barely have time to react as it bears down on you, its massive frame a blur of movement. 
+                As it strikes, you instinctively raise your arms to shield yourself, but its powerful jaws clamp down on your forearm, the jagged teeth slicing through skin and muscle like paper. 
+                Pain explodes through your body, searing and sharp, but you manage to twist away just enough to avoid a fatal bite. 
+                The weight of the rat pushes you backward, and you stumble, trying to regain your balance. 
+                """, """
+                As the monstrous rat locks its glowing red eyes on you, there's a split second where time seems to freeze. 
+                Its hulking, muscular frame ripples as it starts to move, each step deliberate and terrifying. 
+                The sound of its claws scraping against the ground echoes through the narrow, shadowy space you're trapped in. 
+                The smaller rats scatter, leaving a clear path for their leader to close the distance with terrifying speed. 
+                
+                Before you can react, the rat lunges. Its jagged teeth snap together with a sickening crack, narrowly missing your face. 
+                The stench of decay and blood fills the air as its hot, foamy breath washes over you. 
+                With a powerful swipe of its clawed paw, it sends you sprawling to the ground. 
+                The impact knocks the wind out of you, and you can feel its claws raking across your chest, tearing through your clothes and skin with ease. 
+                Gasping for air, you struggle to get up, but the rat is already on you. 
+                Its massive jaws clamp down on your shoulder, the sharp fangs easily piercing flesh and muscle. 
+                The pain is excruciating, but what makes it worse is the gnashing—those discolored, broken teeth grind through bone with a sickening crunch. 
+                
+                You scream, but the sound is drowned out by the rat's guttural growl as it shakes you violently, like a predator toying with its prey. 
+                It doesn\'t end quickly. The rat toys with you, pinning you down with its immense weight. 
+                Its claws dig into your sides as it bites down again, this time on your neck. 
+                The world begins to blur as your body weakens, the blood loss too much. 
+                The last thing you hear is the sound of its teeth grinding together, the wet, tearing sound of flesh, and the awful, thick froth of saliva mixing with your blood. 
+                Then, darkness.
+                """, """
+                The rat lets out a guttural growl, then charges with terrifying speed, its massive form closing the distance faster than you expected. 
+                You duck behind a nearby clothing rack, shoving it toward the rat as it barrels forward. 
+                The rack crashes into its legs, but the beast barely falters, smashing through it with raw power. 
+                You need something stronger to stop its momentum. 
+                You run into a nearby sporting goods store, the rat still in pursuit. 
+                Inside the store, you grab a sturdy aluminum baseball bat from a display. 
+                It\'s not much, but it\'s better than nothing. 
+                
+                You also spot cans of spray paint and lighter fluid on a shelf nearby, potential weapons if used correctly. 
+                You stuff a few into your pockets, hoping to use them later. 
+                The rat bursts into the store, its massive bulk knocking over shelves as it forces its way inside. 
+                You swing the bat with all your strength, smashing it across the rat\'s face. 
+                The impact sends a jolt through your arms, but the rat barely flinches, only snarling louder. 
+                It swipes at you with one massive paw, and you barely manage to dodge, the claws missing your torso by inches. 
+                Remembering the lighter fluid, you pull a canister from your pocket and spray it at the rat\'s face, hoping to blind or disorient it. 
+                The rat screeches as the chemical splashes into its eyes, causing it to thrash around wildly. 
+                You take the opportunity to grab a lighter from a nearby counter, igniting the fluid with a quick flick. 
+                
+                The rat\'s fur catches fire, and the beast lets out an ear-piercing howl of pain. 
+                Flames race across its back and sides, causing it to thrash even more violently. 
+                Its eyes are wild with rage, and it snaps at you as you get closer. 
+                You wait for the right moment, when it turns its head, exposing its side. 
+                With all the strength you can muster, you bring the bat down to the back of its neck. 
+                There\'s a crack as the blow connects, and the rat collapses, its legs giving out beneath it. 
+                Breathing heavily, you drop the bat and stagger back, the adrenaline still coursing through your veins. 
+                """);
+    }
+}
