+++ Adventure_Game/Ratpack.java
@@ -1,0 +1,70 @@
+package monster;
+
+public class Ratpack extends Monster {
+    public Ratpack() {
+        super(2, """
+                A pack of rabid rats surges through the mall like a chaotic, writhing mass of fur, teeth, and claws. 
+                Their bodies, numbering in the dozens or even hundreds, are mottled with patches of mangy fur, tufts missing where disease has ravaged their once sleek coats. 
+                The rats vary in size, some as large as small cats, with their muscular forms twitching and jerking with unnatural energy. 
+                Their thin, whip-like tails lash the ground behind them, dragging through the dirt and grime, adding to the noise of their frenzied scurrying. 
+                Foam bubbles and drips from the corners of their mouths, a clear sign of the rabies coursing through their bodies. 
+                Their sharp, yellowed teeth are constantly exposed, gnashing together with an audible grinding sound. 
+                
+                The smell of the pack is overpowering. A thick, nauseating stench of decay, rot, and disease. 
+                The air around them is thick with the odor of sweat, blood, and festering wounds, combined with the sickly-sweet smell of their frothing mouths. 
+                Their movement is erratic and terrifyingly fast. 
+                They dart forward in a jerking, frantic fashion, never moving in a straight line but instead twisting and turning unpredictably. 
+                As they move the sound is relentless, high-pitched squeals, the gnashing of teeth, the scratching of claws, all of it amplified by the sheer number of bodies moving as one. 
+                By each second, they\'re inching closer to you.  
+                """, """
+                As the pack of rats surges toward you, the adrenaline kicks in, and you make a desperate dash to escape. 
+                You can hear the relentless squeals and the scratching of their claws on the ground behind you, growing louder, closer. 
+                Just as you reach a narrow corridor, the first of them lunges at your legs, a flash of mangy fur and teeth. 
+                You kick out instinctively, but one of them manages to sink its teeth into your ankle. 
+                The pain is sharp and immediate, like a jagged knife slicing through your flesh. 
+                You stumble, almost falling, but catch yourself against the wall. 
+                Your heart pounds as you feel the rat\'s weight hanging from your ankle, its jaws locked tight, teeth grinding against bone.
+                """, """
+                The pack of rats surges toward you, a wave of gnashing teeth and frantic energy. 
+                Panic floods your system as you turn to run, but their speed is overwhelming. 
+                Before you can get far, the first rat leaps at your leg, its claws digging into your skin. 
+                You feel a sharp, searing pain as its teeth sink deep into your calf. 
+                Your body tenses in shock, but there\'s no time to react. 
+                The rest of the pack follows, swarming you like a tidal wave of fur and frenzy. 
+
+                They claw and bite with savage intensity, their sharp teeth tearing through your clothes and skin, ripping into muscle with a sickening crunch. 
+                Their sheer numbers overwhelm you in seconds. 
+                You try to fight them off, your hands swatting in a desperate attempt to push them away, but they are everywhere, crawling up your arms, biting into your flesh, even latching onto your face. 
+                The sensation is unbearable: a constant, searing pain as dozens of tiny mouths gnaw at you relentlessly. 
+                You can feel their claws raking across your skin, peeling away layers as they bite deeper. 
+                The froth from their mouths smears across your skin as they bite down, infecting every wound with rabies. 
+                Your vision blurs as they tear at your eyes, and the metallic taste of blood fills your mouth as they bite into your lips, gnashing teeth against bone. 
+
+                Your screams are lost beneath the deafening squeals of the pack. 
+                The rats chew and burrow into every inch of your body, gnawing through nerves and tendons, reducing your movements to feeble twitches. 
+                The pain fades to numbness as your body succumbs to shock, but the image of them. 
+                Those yellowed teeth, those wild, rabid eyes, burns into your mind as the last thing you see. 
+                Darkness claims you, but the gnawing continues long after. 
+                """, """
+                First, you run toward the nearest store with a wide entrance to put some distance between you and the approaching pack. 
+                The rats are fast, but you are faster, for now. 
+                You slam the glass door shut behind you, buying yourself a few precious seconds. 
+                You quickly grab whatever is near the door, clothes racks, a bench, or even mannequins, and shove them against the entrance to create an improvised barrier. 
+                It\'s not perfect, but it will slow them down. 
+                The rats slam into the glass, claws scratching, their squeals filling the air as they push against the door, but you know they\'ll break through soon. 
+                The first rat bursts through the door, squeezing through the crack it\'s gnawed open. 
+                Its eyes are wild, saliva dripping from its jaws as it scurries forward, claws clicking on the tiled floor. 
+                More follow, pushing their way inside. 
+                
+                Desperate, you back up to the store\'s counter, spotting a fire extinguisher mounted on the wall. 
+                Perfect. You grab it and rip it off the mount, holding it up like a shield. 
+                As the rats close in, you spray the thick foam at the swarm, the blast disorienting them. 
+                The cold, pressurized cloud blinds them, sending them into a frenzy of confusion. 
+                Using the extinguisher as a blunt weapon, you swing it down with all your strength, crushing several of the rats under its weight. 
+                Each swing sends more of them flying, their bodies breaking under the force of the heavy metal canister. 
+                With the swarm now thinned out, the remaining rats are struggling to regroup. 
+                You take the chance to make a break for it, vaulting over the counter and rushing out of the store. 
+                You don\'t stop running, knowing that while you\'ve dealt with the immediate threat, the mall is still crawling with danger. 
+                """);
+    }
+}
