+++ Adventure_Game/Zombie.java
@@ -1,0 +1,51 @@
+package monster;
+
+public class Zombie extends Monster {
+
+    public Zombie(){
+        super(1,"""
+            Close by you hear footsteps approaching slowly and a low guttural moan intermittently. 
+            Curious on what it may be you sneak up closer to get a look. 
+            What you see is a person wearing a hoodie and jeans, tattered and covered in grime and... blood? 
+            The face seems almost lifeless with its jaw hanging slightly agape revealing just a few teeth remaining, smeared with dark, coagulated blood. 
+            Its skin appears to be sickly greenish grey and stretched tightly over its skeletal features. Clumps of hair cling to its scalp, greasy and matted with dried blood and dirt. 
+            
+            The eyes are glassy and vacant, staring ahead into nothing. 
+            Its movements are slow and jerky, dragging one foot slightly behind. 
+            Taken aback of what you just saw, you didn\'t notice the broken glass bottle beside you, accidentally kicking it. 
+            The zombie\'s gaze flicked instantly and started moving towards you.  
+            """, """
+            As the zombie charges toward you with surprising speed, you barely manage to take a step back before it lunges, its filthy hand slamming into your chest with unexpected force. 
+            The impact knocks the wind out of you, sending you stumbling backward. 
+            Your ribs ache from the hit, the pain sharp and immediate. 
+            """, """
+            As soon as the sound of the bottle shattering echoes, the zombie\'s vacant, glassy eyes lock onto you with sudden intensity. 
+            It lets out a raspy, guttural growl, dragging its stiff leg faster now, lurching in your direction. 
+            Panic floods through your body, your mind racing for a way out, but the creature is relentless, closing the distance quickly despite its awkward gait. 
+            Before you can react, it's upon you. Its decaying hands, caked with filth, grab your arm with an iron grip far stronger than you expected. 
+            The stench of death floods your senses as it pulls you in closer, jaw dropping open, revealing its remaining teeth stained with dried blood. 
+            You struggle, but its dead weight overpowers you as you tumble to the ground. 
+            The zombie\'s face inches closer to yours, its breath a mix of decay and rot. 
+            
+            With terrifying strength, it forces your head to the side, exposing your neck. 
+            Then, with a sickening crunch, its jagged teeth sink into your flesh, tearing into muscle and tendon as blood pours from the wound. 
+            The pain is excruciating, but it quickly dulls as shock takes over. 
+            You feel the warmth of your own blood soaking your clothes, your heartbeat hammering in your ears. 
+            The world blurs, your vision dimming as the monster feasts hungrily. 
+            The last thing you hear before everything goes black is the wet, grisly sound of it devouring your flesh, your body going cold in its grip. 
+            """, """
+            You run around trying to find a weapon to help you fight it, spotting a car, you rush towards it, heart pounding. 
+            You yank open the driver\'s side door. 
+            With the door wide open, you step inside. 
+            The zombie lurches closer, its vacant eyes fixated on you. 
+            As the zombie draws near, its outstretched arms reaching for you, you suddenly swing the door shut with all your might. 
+            The metal slams against the zombie with a bone-crushing force, the sickening sound of flesh meeting steel echoing through the parking lot. 
+            
+            You aim for the side of its head, hoping to catch it off guard. 
+            The door collides with the zombie\'s head, caving it in and effectively neutralizing the threat. 
+            The creature falls limp, its body crumpling against the asphalt, blood pooling beneath it. 
+            You step out of the car, breathless and adrenaline-fueled, scanning the lot for other zombies that may have been attracted by the noise. 
+            """);
+        
+    } 
+}
