--- Dog.java
@@ -1,58 +1,0 @@
-package monster;
-
-public class Dog extends Monster {
-    public Dog() {
-        super(2, """
-                A big muscular canine, about three feet in height at the shoulder, with a broad chest and sturdy legs. 
-                The fur is jet-black and unkept, clumped in places. 
-                The dog\'s eyes are wide and bloodshot, glowing faintly red in the shadows. 
-                Its movement is erratic, jerking as if it\'s in pain or confused. 
-                Foam drips from its partially open maw, where its sharp, yellowed teeth are bared in a constant snarl, lips pulled back in a grotesque display of aggression. 
-                
-                The frothy saliva clings to the fur around its mouth, and the constant growling emanating from its throat sounds deep, and animalistic. 
-                It appears the dog has caught your sent, and its movement shifted to a slow, stalking manner. 
-                Its ears are pinned back, every step deliberate and tense, as if it\'s ready to pounce at any second. 
-                The fur on its back stands erect, a ridge of bristling black that runs from the back of its neck to the base of its tail, signaling its heightened agitation. 
-                Its breathing is ragged and heavy. 
-                """, """
-                The dog lunges at you, and this time, you manage to react just in time. 
-                With adrenaline surging through your veins, you twist to the side, narrowly avoiding its teeth sinking into your throat. 
-                Instead, the beast's powerful jaws clamp down on your arm. 
-                Pain explodes through you as its sharp teeth pierce through muscle, the force of its bite bruising bone. 
-                Blood pours from the wound. 
-                """, """
-                The dog locks onto you with its glowing, bloodshot eyes, narrowing them as it shifts into a low, predatory stance. 
-                The growl deepens, reverberating through the air as it begins its slow, deliberate approach. 
-                You freeze, the terror taking hold as the animal\'s intent becomes clear. 
-                Each of its steps feels like a countdown. 
-                
-                Without warning, the dog lunges. The speed is startling, a blur of black fur and bared teeth, its mouth opening wider as it propels itself forward. 
-                You barely have time to react before its powerful jaws clamp onto your leg, the impact knocking you to the ground. 
-                The pain is immediate and blinding as the sharp teeth tear through flesh and muscle with savage precision. 
-                Its grip is unyielding, and it begins to shake its head violently, the foam from its mouth splattering in all directions, mixing with your blood. 
-                You try to fight back, instinctively kicking, clawing, but it\'s futile. 
-                
-                The dog is too strong. It releases your leg for a brief moment, only to shift its attention to your throat. 
-                It moves with terrifying accuracy, teeth sinking into the vulnerable spot between your neck and shoulder. 
-                The pressure of the bite crushes your windpipe, cutting off any scream, any plea. 
-                Blood fills your mouth as you choke, the metallic taste overwhelming. 
-                The last thing you feel is the animal\'s hot, ragged breath against your skin as it tightens its jaws, severing the final threads of life, the world fading to black as its growling fades into a distant echo. 
-                """, """
-                Beneath your feet, you find a small rusty iron pipe. It isn't much but it is all you have. 
-                You pick it up and grip it tightly. The dog\'s bloodshot eyes were wild, locked on you with an unnatural, terrifying intensity. 
-                Every muscle in its body was tense, ready to explode into motion. 
-                Suddenly, it lunged. 
-                You barely had time to react as its massive body launched forward, snarling teeth bared, aiming for your throat. 
-                In a split second, you sidestepped, swinging the pipe with every ounce of strength you had. 
-                The metal connected with its ribcage in a sickening thud. 
-                You could feel the impact reverberate through your arms, as though you\'d struck a wall of muscle and bone. 
-                
-                The dog let out a yelp, an unnatural, choking sound. It stumbled mid-air, crashing hard into the floor as the blow rattled its body. 
-                Its legs twitching, struggling to rise, something had broken inside. 
-                Its ribs had shattered under the force of the strike, and you knew you had likely punctured a lung. 
-                It gasped, a wet, ragged sound, as it writhed on the floor. 
-                The growls turned into weak whimpers before finally going silent. 
-                The dog\'s body went limp, its chest rising and falling in short, shallow breaths before stopping altogether. 
-                """);
-        }
-}
