--- Spider.java
@@ -1,61 +1,0 @@
-package monster;
-
-public class Spider extends Monster {
-    public Spider() {
-        super(1, """
-                Even though you feel more at ease around your campfire. 
-                The pitch-black surroundings give an eerie feeling of discomfort. 
-                During the night it's hard to fall asleep, it feels as if the forest has woken up. 
-                It sounds as if a bunch of different animals are lurking in the shadows. 
-                While trying to block out all the noice, you turn to lay on your back, counting the leaves in the tree above, you see something staring right at you. 
-                This giant black spider, towering at almost four feet tall. 
-                
-                Its body is divided into two distinct sections: a broad, rounded abdomen and a smaller, more compact head. 
-                The abdomen is smooth and glossy, reflecting the light with an almost oily sheen. 
-                It pulses slightly as the spider breathes. 
-                Its eight legs are covered in a dense array of tiny hairs that seem to absorb rather than reflect light. 
-                Each leg is segmented into multiple joints that bend in unnatural angles, with sharp and hooked claws at the end of each leg.  
-                The head seems armored with a hardened exoskeleton. 
-                Eight jet-black eyes in two rows, with the top four larger than the others, glimmer like obsidian. 
-                Two menacing fangs curve outward, gleaming and dripping with thick venom. 
-                """, """
-                As the giant spider launches itself from the tree, its massive form slams into you with bone-crushing force, sending you sprawling across the ground. 
-                The impact knocks the wind out of you, and a sharp pain flares through your ribs, but somehow, you're still alive. 
-                """, """
-                Before you even realize it, the massive spider launches itself from the tree above, descending like a living nightmare. 
-                In an instant, its hulking form crashes down, the impact driving the air from your lungs. 
-                The speed at which it moves is terrifying, its long, hooked legs closing the distance between you in the blink of an eye. 
-
-                One of those legs' slams into your chest, pinning you with immense pressure. 
-                The hooked claw at the tip pierces your skin just enough to draw blood, but it\'s the weight that\'s suffocating. 
-                You struggle to breathe, your limbs useless as panic grips you. 
-                All you can do is stare in horror at the glistening black eyes fixed on you, unblinking, like eight lifeless windows to the void. 
-
-                Its fangs flash in the dim light as it rears back slightly, preparing to strike. 
-                In a blur of motion, it lunges, burying both of its venomous fangs deep into your shoulder. 
-                Agonizing pain shoots through your body, followed by an intense burning sensation as the venom floods your system. 
-                Your muscles seize up almost instantly, your body going rigid as the paralyzing toxin takes over. 
-
-                As your consciousness begins to fade, the last thing you feel is the spider dragging you toward the shadows, its monstrous legs effortlessly carrying you into the darkness. 
-                The forest swallows the scene, and you vanish, becoming just another part of the eerie silence. 
-                """, """
-                The massive spider launches towards you with its legs outstretched and ready to strike. 
-                Your heart races, instinct kicking in. 
-                The spider\'s monstrous frame is coming straight for you, its massive fangs poised to bite. 
-                The sheer size of it sends a surge of adrenaline through you as you realize you have only moments to act. 
-                With a desperate rush, you grab a branch from the campfire. 
-                
-                As the spider reaches the apex of its leap, you swing the flaming torch upward, aiming for the monstrous creature. 
-                The fire crackles, sending up sparks that dance in the air. 
-                With a primal roar, you thrust the torch forward just as the spider lands, its weight crashing down beside you, sending a tremor through the ground. 
-                The spider recoils from the flames, its legs flailing as it instinctively retreats from the heat. 
-                You take advantage of its hesitation, brandishing the torch like a weapon. 
-                You push the torch closer, the flames licking at its legs. 
-                
-                The spider hisses, an unnatural sound that echoes through the night, and you see smoke begin to rise from its bristly fur. 
-                It thrashes wildly, trying to escape the fiery onslaught, but it\'s too late, the heat is overwhelming. 
-                You can see its flesh start to char as the fire bites into its side, causing it to screech in agony. 
-                Breathing heavily, you step back, watching as the once-formidable spider succumbs to the flames.
-                """);
-    }
-}
